+++ .github/workflows/build-validation.yaml
@@ -1,0 +1,18 @@
+name: Gatsby Publish Build Validation
+
+on:
+  pull_request:
+    branches:
+      - dev
+
+jobs:
+  build:
+    runs-on: macos-latest
+    steps:
+      - uses: actions/checkout@v1
+      - uses: enriikke/gatsby-gh-pages-action@v2
+        with:
+          access-token: ${{ secrets.DEPLOY_GATSBY_TO_GH_PAGES }}
+          deploy-branch: gh-pages
+          gatsby-args: --prefix-paths
+          skip-publish: true
