+++ .github/workflows/deploy.yaml
@@ -1,0 +1,17 @@
+name: Gatsby Publish
+
+on:
+  push:
+    branches:
+      - main
+
+jobs:
+  build:
+    runs-on: macos-latest
+    steps:
+      - uses: actions/checkout@v1
+      - uses: enriikke/gatsby-gh-pages-action@v2
+        with:
+          access-token: ${{ secrets.DEPLOY_GATSBY_TO_GH_PAGES }}
+          deploy-branch: gh-pages
+          gatsby-args: --prefix-paths
