+++ .stylelintrc.json
@@ -1,0 +1,7 @@
+{
+    "extends": [
+        "stylelint-config-standard",
+        "stylelint-config-prettier"
+    ]
+    
+  }
