+++ gatsby-config.js
@@ -8,6 +8,7 @@ module.exports = {
     "gatsby-transformer-remark",
     "gatsby-plugin-sharp",
     "gatsby-transformer-sharp",
+    "gatsby-plugin-eslint.js",
     {
       resolve: "gatsby-source-filesystem",
       options: {
